+++ RELEASE_NOTES.md
@@ -1,3 +1,26 @@
-* 0.0.3-alpha - Integrate FSharp.PowerPack.Metadata
-* 0.0.2-alpha - Integrate hosted FSI configuration, SimpleSourceCodeServices, cleanup to SourceCodeServices API
-* 0.0.1-alpha - Rename the FSharp.Compiler.Editor package
+#### 0.0.2-alpha - 
+* Integrate hosted FSI configuration, SimpleSourceCodeServices, cleanup to SourceCodeServices API
+
+#### 0.0.3-alpha - 
+* Integrate FSharp.PowerPack.Metadata as the FSharp* symbol API
+* Renamed Param --> MethodGroupItemParameter and hid record from view, made into an object
+* Renamed Method --> MethodGroupItem and hid record from view, made into an object
+* Renamed Methods --> MethodGroup and hid record from view, made into an object
+* Renamed MethodGroup.Name --> MethodGroup.MethodName
+* Renamed DataTip --> ToolTip consistently across all text
+* Renamed CheckOptions --> ProjectOptions
+* Renamed TypeCheckAnswer --> CheckFileAnswer
+* Renamed UntypedParseInfo --> ParsedFileResults
+* Removed GetCheckOptionsFromScriptRoot member overload in favour of optional argument
+* Renamed GetCheckOptionsFromScriptRoot --> GetProjectOptionsFromScriptRoot
+* Renamed UntypedParse --> ParseFileInProject
+* Renamed TypeCheckSource --> CheckFileInProjectIfReady
+* Added numerous methods to API including CheckFileInProject
+* Added experimental GetBackgroundCheckResultsForFileInProject, GetBackgroundParseResultsForFileInProject
+* Added PartialAssemblySignature to TypeCheckResults/CheckFileResults
+* Added CurrentPartialAssemblySignature to FsiEvaluationSession 
+* Added ParseAndCheckInteraction to FsiEvaluationSession to support intellisense implementation against a script fragment 
+* Added initial testing in tests/service
+* Added ParseAndCheckProject to SourceCodeServices API. This will eventually return "whole project" information such as symbol tables.
+* Added GetDefaultConfiguration to simplify process of configuring FsiEvaluationSession
+* Added TypeStateUpdated event to FsiEvaluationSession